extract argo interactions to own file
