+++ internal/connection/octant.go
@@ -1,7 +1,6 @@
 package connection
 
 import (
-	"bytes"
 	"context"
 	_ "embed"
 	"encoding/json"
@@ -119,114 +118,5 @@ func (oc *OctantConnection) DeleteConnection(ctx context.Context, namespace, con
 		return fmt.Errorf("failed to update configmap %s after deletion: %w", connectionsConfigmapName, err)
 	}
 
-	return nil
-}
-
-func (oc *OctantConnection) getArgoAppStatus(ctx context.Context, namespace, name string) error {
-	// GET APP
-	return nil
-}
-
-func (oc *OctantConnection) pushArgoApp(ctx context.Context, namespace, name string, connection OctantConnectionData) error {
-	// FIXME: Actually wire up the integration name here
-	datadawgIntegration, getDDIntErr := oc.DataDogIntegrationStuff.GetIntegrationByName(ctx, namespace, "datadawg")
-	if getDDIntErr != nil {
-		return getDDIntErr
-	}
-	argoIntegration, getArgoIntErr := oc.ArgoCDIntegrationStuff.GetIntegrationByName(ctx, namespace, "default-argo-integration")
-	if getArgoIntErr != nil {
-		return getArgoIntErr
-	}
-
-	templateData := ArgoTemplateData{
-		AppName:        name,
-		Namespace:      namespace,
-		ConnectionData: connection,
-		TempDDAPIKey:   datadawgIntegration.APIKey,
-		TempDDURL:      datadawgIntegration.DDUrl,
-		// Tells template to manually inject Argo tracking annotations. We only want these for direct sync force push
-		IsArgoSideload: true,
-	}
-
-	appCreateErr := oc.doArgoAppCreation(ctx, templateData, argoIntegration)
-	if appCreateErr != nil {
-		return appCreateErr
-	}
-
-	syncErr := oc.doArgoAppSync(ctx, templateData, argoIntegration, name)
-	if syncErr != nil {
-		return syncErr
-	}
-
-	return nil
-}
-
-func (oc *OctantConnection) doArgoAppSync(ctx context.Context, templateData ArgoTemplateData, argoIntegration *integration.ArgoCDIntegrationData, name string) error {
-	manifests, err := oc.renderSyncManifests(&templateData)
-	if err != nil {
-		return err
-	}
-
-	syncPayload := map[string]any{
-		"revision": "HEAD",
-		"prune":    false,
-		"dryRun":   false,
-		"strategy": map[string]interface{}{
-			"apply": map[string]bool{
-				"force": false,
-			},
-		},
-		"manifests": manifests,
-	}
-	syncPayloadJson, err := json.Marshal(syncPayload)
-	// FIXME: Delete this, I just wanted a convenient place to see the manifests
-	syncPayloadStr := string(syncPayloadJson)
-	if syncPayloadStr == "erresrlkj" {
-		return fmt.Errorf("hehehhhhhh")
-	}
-	if err != nil {
-		return err
-	}
-	syncUrl := fmt.Sprintf("%s/api/v1/applications/%s/sync", argoIntegration.APIUrl, name)
-	syncReq, err := http.NewRequestWithContext(ctx, "POST", syncUrl, bytes.NewReader(syncPayloadJson))
-	if err != nil {
-		return err
-	}
-	syncReq.Header.Set("Content-Type", "application/json")
-	syncReq.Header.Set("Authorization", fmt.Sprintf("Bearer %s", argoIntegration.AccountToken))
-	syncResp, err := oc.httpClient.Do(syncReq)
-	if err != nil {
-		return err
-	}
-	if syncResp.StatusCode != http.StatusOK {
-		return fmt.Errorf("unexpected status code: %d", syncResp.StatusCode)
-	}
-	return nil
-}
-
-func (oc *OctantConnection) doArgoAppCreation(ctx context.Context, templateData ArgoTemplateData, argoIntegration *integration.ArgoCDIntegrationData) error {
-	appJson, err := oc.renderArgoAppManifest(&templateData)
-	if err != nil {
-		return err
-	}
-	createAppUrl := fmt.Sprintf("%s/api/v1/applications", argoIntegration.APIUrl)
-	req, err := http.NewRequestWithContext(ctx, "POST", createAppUrl, bytes.NewReader(appJson))
-	if err != nil {
-		return err
-	}
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Authorization", fmt.Sprintf("Bearer %s", argoIntegration.AccountToken))
-	resp, err := oc.httpClient.Do(req)
-	if err != nil {
-		return err
-	}
-	if resp.StatusCode != http.StatusOK {
-		return fmt.Errorf("unexpected status code: %d", resp.StatusCode)
-	}
-	return nil
-}
-
-func (oc *OctantConnection) deleteArgoApp(ctx context.Context, namespace, name string) error {
-	// DELETE APP
 	return nil
 }
